models/requirements.txt
huggingface_hub
pyyaml

